[fix]: 마인드맵 노드 디자인 수정
- rootNode
- summaryNode
