+++ frontend/src/components/reactFlow/nodes/ui/SummaryNode.tsx
@@ -1,16 +1,12 @@
-import { Modal } from '@/components/common/Modal';
 import NodeHandles from '@/components/reactFlow/nodes/ui/NodeHandles';
-import { Button } from '@/components/ui/button';
-import { DialogClose } from '@/components/ui/Dialog';
-import { useDeleteNode, useNodes, useSetNode } from '@/store/mindMapStore';
+import { useNodes, useSetNode } from '@/store/mindMapStore';
 import { SummaryNodeType } from '@/types/mindMap';
 import { NodeProps } from '@xyflow/react';
-import { GripVertical, Pencil, X } from 'lucide-react';
+import { CornerDownRight, GripVertical, PencilLine } from 'lucide-react';
 
 export default function SummaryNode({ id, data }: NodeProps<SummaryNodeType>) {
   const nodes = useNodes();
   const setNode = useSetNode();
-  const deleteNode = useDeleteNode();
 
   const handleEditClick = () => {
     const currentNode = nodes.find((node) => node.id === id);
@@ -33,47 +29,14 @@ export default function SummaryNode({ id, data }: NodeProps<SummaryNodeType>) {
         <p className="text-[20px] font-semibold">{data.summary}</p>
       </div>
       <div className="opacity-0 group-hover:opacity-100 flex items-center gap-4 transition-opacity duration-300 ease-in-out">
-        <Pencil
+        <PencilLine
           size={20}
           color="#B9B9B7"
           className="cursor-pointer hover:text-black transition-colors z-20"
           onClick={handleEditClick}
         />
+        <CornerDownRight size={20} color="#B9B9B7" />
 
-        <Modal
-          trigger={
-            <X
-              size={20}
-              color="#B9B9B7"
-              className="cursor-pointer hover:text-black transition-colors z-20"
-            />
-          }
-          title="이 노드를 삭제할까요?"
-          description="해당 노드와 모든 하위노드가 함께 삭제돼요"
-          footer={
-            <div className="w-full flex items-center justify-between">
-              <DialogClose asChild>
-                <Button className="px-8" variant="white">
-                  취소하기
-                </Button>
-              </DialogClose>
-              <DialogClose asChild>
-                <Button
-                  className="px-8"
-                  onClick={() => {
-                    deleteNode(id);
-                  }}
-                >
-                  삭제하기
-                </Button>
-              </DialogClose>
-            </div>
-          }
-        >
-          <div className="rounded-xl px-6 py-4 font-bold border-2 border-border-gray">
-            {data.summary}
-          </div>
-        </Modal>
         <div className="dragHandle cursor-grab z-20">
           <GripVertical
             size={20}
